requirements.txt
django
model-bakery
psycopg2-binary
djangorestframework
requests